Add gts and reorganize
