+++ Makefile
@@ -14,7 +14,7 @@ python-build: rust-build
 js-build:
 	wasm-pack build --dev --target nodejs polar
 	rm polar/pkg/.gitignore polar/pkg/package.json
-	cp --recursive polar/pkg/. languages/js/lib/wasm
+	cp --recursive polar/pkg/. languages/js/lib
 	rm -rf polar/pkg
 
 python-test: python-build
