+++ languages/js/.prettierrc.js
@@ -1,0 +1,4 @@
+module.exports = {
+  ...require('gts/.prettierrc.json'),
+  bracketSpacing: true,
+}
