+++ languages/js/lib/polar.d.ts
