+++ languages/js/lib/polar_bg.d.ts
